refactor: make graphviz a package
