+++ packages/remark-graphviz/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@tdev/remark-pdf",
+    "name": "@tdev/remark-graphviz",
     "version": "1.0.0",
     "main": "lib/index.js",
     "types": "lib/remark-pdf.d.ts",
@@ -9,7 +9,7 @@
         "dev": "tsc --watch"
     },
     "dependencies": {
-        "react-pdf": "^9.2.1"
+        "@hpcc-js/wasm-graphviz": "^1.7.0"
     },
     "devDependencies": {
         "vitest": "*",
